refactor : support 패키지 config 코드 이동
